Added user phase control

+++ MCSDK_v6.0.0-Full/MotorControl/MCSDK/MCLib/Any/Inc/speed_torq_ctrl.h
@@ -129,7 +129,7 @@ MC_ControlMode_t STC_GetControlMode(SpeednTorqCtrl_Handle_t *pHandle);
 bool STC_ExecRamp(SpeednTorqCtrl_Handle_t *pHandle, int16_t hTargetFinal, uint32_t hDurationms);
 
 /* Starts the execution of a sin speed. */
-bool STC_ExecSin(SpeednTorqCtrl_Handle_t *pHandle, int16_t hMean, int16_t hAmp,  int16_t hPhase);
+bool STC_ExecSin(SpeednTorqCtrl_Handle_t *pHandle, int16_t hMean, int16_t hAmp,  int32_t hPhase);
 /* It interrupts the execution of any previous ramp command.*/
 void STC_StopRamp(SpeednTorqCtrl_Handle_t *pHandle);
